Fix swell-wnd compilation (macOS)

+++ WDL/swell/swell-wnd.mm
@@ -874,7 +874,7 @@ STANDARD_CONTROL_NEEDSDISPLAY_IMPL( m_lbMode ? "SysListView32_LB" : "SysListView
     if (m_start_item >=0 && (m_fastClickMask&(1<<m_start_subitem)))
     {
       int msg = [theEvent clickCount] > 1 ? NM_DBLCLK : NM_CLICK;
-      NMLISTVIEW nmlv={{(HWND)self,(UINT_PTR)[self tag], msg}, m_start_item, m_start_subitem, 0, 0, 0, {NSPOINT_TO_INTS(pt)}, };
+      NMLISTVIEW nmlv={{(HWND)self,(UINT_PTR)[self tag], static_cast<UINT>(msg)}, m_start_item, m_start_subitem, 0, 0, 0, {NSPOINT_TO_INTS(pt)}, };
       SWELL_ListView_Row *row=m_items->Get(nmlv.iItem);
       if (row)
         nmlv.lParam = row->m_param;
